Rework

+++ src/CreditPaymentService.java
@@ -1,5 +1,5 @@
 public class CreditPaymentService {
-    public int calculate(float creditSum, float yearlyPercentageRate, float creditTerm) {
+    public int calculate(int creditSum, float yearlyPercentageRate, int creditTerm) {
         float monthlyPercentageRate = yearlyPercentageRate / 100 / 12;
         int monthlyPayment = (int) (creditSum * (monthlyPercentageRate +
                 (monthlyPercentageRate / (Math.pow((1 + monthlyPercentageRate), creditTerm) - 1))));
